requirements.txt
transformers
torch
tqdm
python-magic-bin